Gnomes successfully implemented - They still need some holdings, a cultural building of their own, some events?, a custom building placeholder, and a place to call home.. Thinking: Mahakam and Tir Tochair. -Gnomes have a very high learning skill/ +5 learning -They also have a slight intrigue boost/ +2 intrigue
On a side note...
Sorry about the current Gnomish portrait, it's the Druedain from the MEP.. I know, it's not technically a gnome but it's currently my best option as of now. I'll try my hardest to find a way to create a custom Gnomish portrait. But for now it has to suffice. Use your imagination!!! :)