Proposed fix for #106
There's some confusion between "is_ill" and "trait = ill".  The trigger for this is "is_ill = yes", and not "trait = ill".  As such, this encompasses a lot of things, such as pneumonic (see "is_illness = yes" for a list of traits that are considered normal illness).  As such, in order to match the trigger, I changed it so that instead of removing just the "ill" trait it actually cures all normal illness.  Since this is a Druid doing the healing that is presumably intended.

However, that then removes the need for a lot of unnecessary druid actions, such as druid.106 to remove pneumonic, or druid.108 to remove syphilitic.  This needs to be tested though.  For example I have seen that even though the "Cast Heal" action should do "cure_illness = yes" it did not in fact remove Lover's Pox or Great Pox (when it should since at least syphilitic is considered "is_illness = yes").